requirements/dev.txt
alabaster==1.0.0 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
anyio==4.11.0 ; python_version == "3.12"
asgiref==3.11.0 ; python_version == "3.12"
astroid==3.3.11 ; python_version == "3.12"
asttokens==3.0.1 ; python_version == "3.12"
attrs==25.4.0 ; python_version == "3.12"
babel==2.17.0 ; python_version == "3.12"
brotli==1.2.0 ; platform_python_implementation == "CPython" and python_version == "3.12"
brotlicffi==1.1.0.0 ; platform_python_implementation != "CPython" and python_version == "3.12"
cachetools==6.2.2 ; python_version == "3.12"
certifi==2025.11.12 ; python_version == "3.12"
cffi==2.0.0 ; python_version == "3.12"
cfgv==3.5.0 ; python_version == "3.12"
chardet==5.2.0 ; python_version == "3.12"
charset-normalizer==3.4.4 ; python_version == "3.12"
click==8.3.1 ; python_version == "3.12"
colorama==0.4.6 ; python_version == "3.12"
coverage==7.12.0 ; python_version == "3.12"
cryptography==44.0.3 ; python_version == "3.12"
cssselect2==0.8.0 ; python_version == "3.12"
decorator==5.2.1 ; python_version == "3.12"
defusedxml==0.7.1 ; python_version == "3.12"
dill==0.4.0 ; python_version == "3.12"
distlib==0.4.0 ; python_version == "3.12"
django-cas-sso==1.2.9 ; python_version == "3.12"
django-cors-headers==4.9.0 ; python_version == "3.12"
django-debug-toolbar==5.2.0 ; python_version == "3.12"
django-extensions==3.2.3 ; python_version == "3.12"
django-ipware==7.0.1 ; python_version == "3.12"
django-smart-ratelimit[redis]==0.8.11 ; python_version == "3.12"
django-tenants==3.9.0 ; python_version == "3.12"
django-typer[rich]==3.4.0 ; python_version == "3.12"
django-watchfiles==1.4.0 ; python_version == "3.12"
django-weasyprint==2.4.0 ; python_version == "3.12"
django==4.2.26 ; python_version == "3.12"
djangorestframework-camel-case==1.4.2 ; python_version == "3.12"
djangorestframework-simplejwt==5.5.1 ; python_version == "3.12"
djangorestframework==3.16.1 ; python_version == "3.12"
djangosaml2==1.11.1 ; python_version == "3.12"
docutils==0.21.2 ; python_version == "3.12"
drf-nested-routers==0.94.2 ; python_version == "3.12"
drf-spectacular==0.28.0 ; python_version == "3.12"
elementpath==4.8.0 ; python_version == "3.12"
executing==2.2.1 ; python_version == "3.12"
factory-boy==3.3.3 ; python_version == "3.12"
faker==38.2.0 ; python_version == "3.12"
filelock==3.20.0 ; python_version == "3.12"
fonttools[woff]==4.60.1 ; python_version == "3.12"
hiredis==3.3.0 ; python_version == "3.12"
identify==2.6.15 ; python_version == "3.12"
idna==3.11 ; python_version == "3.12"
imagesize==1.4.1 ; python_version == "3.12"
inflection==0.5.1 ; python_version == "3.12"
ipython-pygments-lexers==1.1.1 ; python_version == "3.12"
ipython==9.7.0 ; python_version == "3.12"
isort==6.1.0 ; python_version == "3.12"
jedi==0.19.2 ; python_version == "3.12"
jinja2==3.1.6 ; python_version == "3.12"
jsonschema-specifications==2025.9.1 ; python_version == "3.12"
jsonschema==4.25.1 ; python_version == "3.12"
markdown-it-py==4.0.0 ; python_version == "3.12"
markupsafe==3.0.3 ; python_version == "3.12"
matplotlib-inline==0.2.1 ; python_version == "3.12"
mccabe==0.7.0 ; python_version == "3.12"
mdurl==0.1.2 ; python_version == "3.12"
nodeenv==1.9.1 ; python_version == "3.12"
packaging==25.0 ; python_version == "3.12"
parameterized==0.9.0 ; python_version == "3.12"
parso==0.8.5 ; python_version == "3.12"
pexpect==4.9.0 ; sys_platform != "win32" and sys_platform != "emscripten" and python_version == "3.12"
pillow==12.0.0 ; python_version == "3.12"
platformdirs==4.5.0 ; python_version == "3.12"
pluggy==1.6.0 ; python_version == "3.12"
pre-commit==4.4.0 ; python_version == "3.12"
prompt-toolkit==3.0.52 ; python_version == "3.12"
psycopg2-binary==2.9.11 ; python_version == "3.12"
ptyprocess==0.7.0 ; sys_platform != "win32" and sys_platform != "emscripten" and python_version == "3.12"
pure-eval==0.2.3 ; python_version == "3.12"
pycodestyle==2.14.0 ; python_version == "3.12"
pycparser==2.23 ; implementation_name != "PyPy" and python_version == "3.12"
pydantic-core==2.41.5 ; python_version == "3.12"
pydantic==2.12.4 ; python_version == "3.12"
pydyf==0.11.0 ; python_version == "3.12"
pygments==2.19.2 ; python_version == "3.12"
pyjwt==2.10.1 ; python_version == "3.12"
pylint==3.3.9 ; python_version == "3.12"
pyopenssl==24.3.0 ; python_version == "3.12"
pyphen==0.17.2 ; python_version == "3.12"
pyproject-api==1.10.0 ; python_version == "3.12"
pysaml2==7.5.0 ; python_version == "3.12"
python-dateutil==2.9.0.post0 ; python_version == "3.12"
python-ipware==3.0.0 ; python_version == "3.12"
python-stdnum==2.1 ; python_version == "3.12"
pytz==2025.2 ; python_version == "3.12"
pyyaml==6.0.3 ; python_version == "3.12"
redis==7.1.0 ; python_version == "3.12"
redis[hiredis]==7.1.0 ; python_version == "3.12"
referencing==0.37.0 ; python_version == "3.12"
requests==2.32.5 ; python_version == "3.12"
rich==14.2.0 ; python_version == "3.12"
roman-numerals-py==3.1.0 ; python_version == "3.12"
rpds-py==0.29.0 ; python_version == "3.12"
ruff==0.11.13 ; python_version == "3.12"
sentry-sdk==2.45.0 ; python_version == "3.12"
shellingham==1.5.4 ; python_version == "3.12"
six==1.17.0 ; python_version == "3.12"
sniffio==1.3.1 ; python_version == "3.12"
snowballstemmer==3.0.1 ; python_version == "3.12"
sphinx==8.2.3 ; python_version == "3.12"
sphinxcontrib-applehelp==2.0.0 ; python_version == "3.12"
sphinxcontrib-devhelp==2.0.0 ; python_version == "3.12"
sphinxcontrib-htmlhelp==2.1.0 ; python_version == "3.12"
sphinxcontrib-jsmath==1.0.1 ; python_version == "3.12"
sphinxcontrib-qthelp==2.0.0 ; python_version == "3.12"
sphinxcontrib-serializinghtml==2.0.0 ; python_version == "3.12"
sqlparse==0.5.3 ; python_version == "3.12"
stack-data==0.6.3 ; python_version == "3.12"
tinycss2==1.5.0 ; python_version == "3.12"
tinyhtml5==2.0.0 ; python_version == "3.12"
tomlkit==0.13.3 ; python_version == "3.12"
tox==4.32.0 ; python_version == "3.12"
traitlets==5.14.3 ; python_version == "3.12"
typer-slim==0.20.0 ; python_version == "3.12"
typing-extensions==4.15.0 ; python_version == "3.12"
typing-inspection==0.4.2 ; python_version == "3.12"
tzdata==2025.2 ; python_version == "3.12"
uritemplate==4.2.0 ; python_version == "3.12"
urllib3==2.5.0 ; python_version == "3.12"
virtualenv==20.35.4 ; python_version == "3.12"
watchfiles==1.1.1 ; python_version == "3.12"
wcwidth==0.2.14 ; python_version == "3.12"
weasyprint==66.0 ; python_version == "3.12"
webencodings==0.5.1 ; python_version == "3.12"
xmlschema==2.5.1 ; python_version == "3.12"
zopfli==0.4.0 ; python_version == "3.12"
zxcvbn==4.5.0 ; python_version == "3.12"
